Move user selectors from services to `selectors.py`
