+++ apps/users/services/base.py
@@ -1,14 +1,13 @@
 import logging
-from typing import Optional
 
 from allauth.account.models import EmailAddress
 from allauth.socialaccount.models import SocialAccount
 from django.core.exceptions import ValidationError
 from django.db import connection, transaction
-from django.db.models import Count
-from django.db.models.query import QuerySet
 
 from users.models import Profile, User
+
+from ..selectors import get_pending_email_address
 
 
 logger = logging.getLogger("default_logger")
@@ -30,34 +29,6 @@ def activate_user(user):
     EmailAddress.objects.create(
         user=user, email=user.email, verified=True, primary=True
     )
-
-
-def get_all_users() -> QuerySet[User]:
-    return User.objects.all()
-
-
-def get_user_by_id(user_id: int) -> User:
-    return User.objects.get(id=user_id)
-
-
-def get_user_by_username(username: str) -> User:
-    return User.objects.get(username=username)
-
-
-def find_user_profiles_with_subscribers() -> QuerySet[Profile]:
-    """Returns a queryset of profiles belonging to users that have at
-    least 1 subscriber.
-    """
-    return Profile.objects.annotate(subscribers_count=Count("subscribers")).filter(
-        subscribers_count__gt=0
-    )
-
-
-def get_all_supscriptions_of_user(user: User) -> list[str]:
-    """Returns a list of usernames of all authors the specified user is
-    subscribed to.
-    """
-    return [p.user.username for p in user.subscribed_profiles.all()]
 
 
 def toggle_user_supscription(user: User, author: User) -> None:
@@ -109,13 +80,6 @@ def delete_pending_email_address(user: User) -> None:
         logger.warning(
             "Attempt of removing non-existent EmailAddress for User(id=%s)", user.id
         )
-
-
-def get_pending_email_address(user: User) -> Optional[EmailAddress]:
-    try:
-        return EmailAddress.objects.get(user=user, primary=False, verified=False)
-    except EmailAddress.DoesNotExist:
-        return None
 
 
 def delete_social_accounts_with_email(email: str) -> None:
